#  Change when RecipientCSV 'get_rows' method gets called