--- a/notifications_utils/clients/redis/__init__.py
+++ b/notifications_utils/clients/redis/__init__.py
@@ -7,7 +7,3 @@ def daily_limit_cache_key(service_id):
 
 def rate_limit_cache_key(service_id, api_key_type):
     return "{}-{}".format(str(service_id), api_key_type)
-
-
-def sms_billable_units_cache_key(service_id):
-    return "{}-sms_billable_units".format(service_id)