The RecipientCSV had a `rows` attribute which generated all the rows when the class was instantiated. This commit changes `rows` to be a property so that the `get_rows` generator only gets called when it is needed.

*Commits are appearing in the wrong order below (due to rebasing) but are in the right order if you check out the branch locally*