@quis 2019-04-18
I think the idea behind this is solid, but it could be refactored a bit: c62adba