@klssmith 2019-04-23
Thanks for the suggestions - I've made the changes to the tests, and cherry-picked your refactoring commit into this branch