@klssmith 2019-04-24
I've tried this out with a 50,000 row CSV file to compare the timings of doing things the old way (using the code on the master branch and calling `.rows`) vs the new way (the code on this branch, using `.get_rows()`).

| | Old way | New way |
--- | --- | --- 
| Average time taken to call `.rows` / `.get_rows()`| 3.999 seconds | 0.033 seconds |
| Average time taken to iterate through 50,000 rows | 4.090 seconds | 3.880 seconds |

Iterating through lots of rows is (very) slightly faster, but the main change would be that we start processing the rows a couple of seconds sooner in notifications-api.